+++ Binary_Tree/Readme.md
@@ -1,3 +1,4 @@
 Binary Tree Algorithms
     - PreOrder Traversal Algorithm
     - PostOrder Traversal Algorithm
+    - Inorder Traversal
